new sources

+++ Telegram/SourceFiles/window/window_controller.cpp
@@ -30,6 +30,7 @@ https://github.com/telegramdesktop/tdesktop/blob/master/LEGAL
 #include "window/themes/window_theme.h"
 #include "window/themes/window_theme_editor.h"
 #include "ui/boxes/confirm_box.h"
+#include "data/data_peer.h"
 #include "mainwindow.h"
 #include "apiwrap.h" // ApiWrap::acceptTerms.
 #include "facades.h"
@@ -40,8 +41,19 @@ https://github.com/telegramdesktop/tdesktop/blob/master/LEGAL
 
 namespace Window {
 
-Controller::Controller()
-: _widget(this)
+Controller::Controller() : Controller(CreateArgs{}) {
+}
+
+Controller::Controller(
+	not_null<PeerData*> singlePeer,
+	MsgId showAtMsgId)
+: Controller(CreateArgs{ singlePeer.get() }) {
+	showAccount(&singlePeer->account(), showAtMsgId);
+}
+
+Controller::Controller(CreateArgs &&args)
+: _singlePeer(args.singlePeer)
+, _widget(this)
 , _adaptive(std::make_unique<Adaptive>())
 , _isActiveTimer([=] { updateIsActive(); }) {
 	_widget.init();
@@ -54,6 +66,14 @@ Controller::~Controller() {
 }
 
 void Controller::showAccount(not_null<Main::Account*> account) {
+	showAccount(account, ShowAtUnreadMsgId);
+}
+
+void Controller::showAccount(
+		not_null<Main::Account*> account,
+		MsgId singlePeerShowAtMsgId) {
+	Expects(isPrimary() || &_singlePeer->account() == account);
+
 	const auto prevSessionUniqueId = (_account && _account->sessionExists())
 		? _account->session().uniqueId()
 		: 0;
@@ -76,24 +96,18 @@ void Controller::showAccount(not_null<Main::Account*> account) {
 
 	_account->sessionValue(
 	) | rpl::start_with_next([=](Main::Session *session) {
+		if (!isPrimary() && (&_singlePeer->session() != session)) {
+			Core::App().closeWindow(this);
+			return;
+		}
 		const auto was = base::take(_sessionController);
 		_sessionController = session
 			? std::make_unique<SessionController>(session, this)
 			: nullptr;
-		if (_sessionController) {
-			_sessionController->filtersMenuChanged(
-			) | rpl::start_with_next([=] {
-				sideBarChanged();
-			}, _sessionController->lifetime());
-		}
-		if (session && session->settings().dialogsFiltersEnabled()) {
-			_sessionController->toggleFiltersMenu(true);
-		} else {
-			sideBarChanged();
-		}
+		setupSideBar();
 		_widget.updateWindowIcon();
 		if (session) {
-			setupMain();
+			setupMain(singlePeerShowAtMsgId);
 
 			session->updates().isIdleValue(
 			) | rpl::filter([=](bool idle) {
@@ -107,6 +121,10 @@ void Controller::showAccount(not_null<Main::Account*> account) {
 				checkLockByTerms();
 				_widget.updateGlobalMenu();
 			}, _sessionController->lifetime());
+
+			widget()->setInnerFocus();
+
+			session->updates().updateOnline(crl::now());
 		} else {
 			setupIntro();
 			_widget.updateGlobalMenu();
@@ -114,6 +132,30 @@ void Controller::showAccount(not_null<Main::Account*> account) {
 
 		crl::on_main(updateOnlineOfPrevSesssion);
 	}, _accountLifetime);
+}
+
+PeerData *Controller::singlePeer() const {
+	return _singlePeer;
+}
+
+void Controller::setupSideBar() {
+	if (!isPrimary()) {
+		return;
+	}
+	if (!_sessionController) {
+		sideBarChanged();
+		return;
+	}
+	_sessionController->filtersMenuChanged(
+	) | rpl::start_with_next([=] {
+		sideBarChanged();
+	}, _sessionController->lifetime());
+
+	if (_sessionController->session().settings().dialogsFiltersEnabled()) {
+		_sessionController->toggleFiltersMenu(true);
+	} else {
+		sideBarChanged();
+	}
 }
 
 void Controller::checkLockByTerms() {
@@ -126,7 +168,7 @@ void Controller::checkLockByTerms() {
 		}
 		return;
 	}
-	Ui::hideSettingsAndLayer(anim::type::instant);
+	hideSettingsAndLayer(anim::type::instant);
 	const auto box = show(Box<TermsBox>(
 		*data,
 		tr::lng_terms_agree(),
@@ -191,21 +233,26 @@ void Controller::showTermsDelete() {
 		if (const auto session = account().maybeSession()) {
 			session->termsDeleteNow();
 		} else {
-			Ui::hideLayer();
+			hideLayer();
 		}
 	};
 	show(
-		Box<Ui::ConfirmBox>(
-			tr::lng_terms_delete_warning(tr::now),
-			tr::lng_terms_delete_now(tr::now),
-			st::attentionBoxButton,
-			deleteByTerms),
+		Ui::MakeConfirmBox({
+			.text = tr::lng_terms_delete_warning(),
+			.confirmed = deleteByTerms,
+			.confirmText = tr::lng_terms_delete_now(),
+			.confirmStyle = &st::attentionBoxButton,
+		}),
 		Ui::LayerOption::KeepOther);
 }
 
 void Controller::finishFirstShow() {
 	_widget.finishFirstShow();
 	checkThemeEditor();
+}
+
+Main::Session *Controller::maybeSession() const {
+	return _account ? _account->maybeSession() : nullptr;
 }
 
 bool Controller::locked() const {
@@ -243,10 +290,10 @@ void Controller::setupIntro() {
 		: Intro::EnterPoint::Start);
 }
 
-void Controller::setupMain() {
+void Controller::setupMain(MsgId singlePeerShowAtMsgId) {
 	Expects(_sessionController != nullptr);
 
-	_widget.setupMain();
+	_widget.setupMain(singlePeerShowAtMsgId);
 
 	if (const auto id = Ui::Emoji::NeedToSwitchBackToId()) {
 		Ui::Emoji::LoadAndSwitchTo(&_sessionController->session(), id);
@@ -285,6 +332,14 @@ void Controller::showBox(
 
 void Controller::showRightColumn(object_ptr<TWidget> widget) {
 	_widget.showRightColumn(std::move(widget));
+}
+
+void Controller::hideLayer(anim::type animated) {
+	_widget.ui_showBox({ nullptr }, Ui::LayerOption::CloseOther, animated);
+}
+
+void Controller::hideSettingsAndLayer(anim::type animated) {
+	_widget.ui_hideSettingsAndLayer(animated);
 }
 
 void Controller::sideBarChanged() {
@@ -334,7 +389,17 @@ void Controller::preventOrInvoke(Fn<void()> &&callback) {
 
 void Controller::invokeForSessionController(
 		not_null<Main::Account*> account,
+		PeerData *singlePeer,
 		Fn<void(not_null<SessionController*>)> &&callback) {
+	const auto separateWindow = singlePeer
+		? Core::App().separateWindowForPeer(singlePeer)
+		: nullptr;
+	const auto separateSession = separateWindow
+		? separateWindow->sessionController()
+		: nullptr;
+	if (separateSession) {
+		return callback(separateSession);
+	}
 	_account->domain().activate(std::move(account));
 	if (_sessionController) {
 		callback(_sessionController.get());
@@ -342,11 +407,9 @@ void Controller::invokeForSessionController(
 }
 
 QPoint Controller::getPointForCallPanelCenter() const {
-	Expects(_widget.windowHandle() != nullptr);
-
 	return _widget.isActive()
 		? _widget.geometry().center()
-		: _widget.windowHandle()->screen()->geometry().center();
+		: _widget.screen()->geometry().center();
 }
 
 void Controller::showLogoutConfirmation() {
@@ -356,26 +419,20 @@ void Controller::showLogoutConfirmation() {
 		? &sessionController()->session().account()
 		: nullptr;
 	const auto weak = base::make_weak(account);
-	const auto callback = [=] {
-		if (account && !weak) {
-			return;
+	const auto callback = [=](Fn<void()> close) {
+		if (!account || weak) {
+			Core::App().logoutWithChecks(account);
 		}
-		if (account
-			&& account->sessionExists()
-			&& Core::App().exportManager().inProgress(&account->session())) {
-			Ui::hideLayer();
-			Core::App().exportManager().stopWithConfirmation([=] {
-				Core::App().logout(account);
-			});
-		} else {
-			Core::App().logout(account);
+		if (close) {
+			close();
 		}
 	};
-	show(Box<Ui::ConfirmBox>(
-		tr::lng_sure_logout(tr::now),
-		tr::lng_settings_logout(tr::now),
-		st::attentionBoxButton,
-		callback));
+	show(Ui::MakeConfirmBox({
+		.text = tr::lng_sure_logout(),
+		.confirmed = callback,
+		.confirmText = tr::lng_settings_logout(),
+		.confirmStyle = &st::attentionBoxButton,
+	}));
 }
 
 Window::Adaptive &Controller::adaptive() const {
